\name{read.survey-package}
\alias{read.survey-package}
\alias{read.survey}
\docType{package}
\title{
Read Surveymonkey file
}
\description{
Prepare data from Surveymonkey xlsx or csv file for analysis.
}
\details{
\tabular{ll}{
Package: \tab read.survey\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-02-09\cr
License: \tab MIT\cr
}
read.survey(filename, format = 'xlsx', convert = FALSE) opens Surveymonkey export file.
}
\author{
Dirk Kloosterboer

Maintainer: Dirk Kloosterboer <info@dirkmjk.nl>
}
\references{
}

\keyword{ package }
\seealso{
}
\examples{

}
